pyproject.toml
[project]
name = "rt-games"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision",
    "lpips",
    "piq",
    "numpy",
    "scipy",
    "pillow",
    "opencv-python",
    "pandas",
    "tqdm",
    "imageio[ffmpeg]",
    "timm",
    "scikit-learn",
    "scikit-image",
    "torch-fidelity"
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning"
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

